+++ sdk/storage/azure-storage-file-share/src/test/java/com/azure/storage/file/share/ShareAPITests.groovy
@@ -941,4 +941,20 @@ class ShareAPITests extends APISpec {
         expect:
         shareName == primaryShareClient.getShareName()
     }
+
+    // This tests the policy is in the right place because if it were added per retry, it would be after the credentials and auth would fail because we changed a signed header.
+    def "Per call policy"() {
+        given:
+        primaryShareClient.create()
+
+        def shareClient = shareBuilderHelper(interceptorManager, primaryShareClient.getShareName())
+            .addPolicy(getPerCallVersionPolicy()).buildClient()
+
+        when:
+        def response = shareClient.getPropertiesWithResponse(null, null)
+
+        then:
+        notThrown(ShareStorageException)
+        response.getHeaders().getValue("x-ms-version") == "2017-11-09"
+    }
 }