+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## 12.7.0-beta.2 (Unreleased)
+- Added support to specify whether or not a pipeline policy should be added per call or per retry.
 - Added support for setting access tier on a share through ShareClient.create, ShareClient.setAccessTier.
 - Added support for getting access tier on a share through ShareClient.getProperties, ShareServiceClient.listShares